updated t1_B

+++ src/scripts/t1_prepare_B.sh
@@ -534,48 +534,63 @@ if ${flags_T1_parc}; then
         psubcortonly="PARC${i}psubcortonly"    
         psubcortonly="${!psubcortonly}"
 
-        echo "${parc} parcellation intersection with GM; pcort is -- ${pcort}"
-
         fileIn="${T1path}/T1_parc_${parc}.nii.gz"
         checkisfile ${fileIn}
 
-        fileOut="${T1path}/T1_parc_${parc}_dil.nii.gz"
+        if [[ "${psubcortonly}" -eq 0 ]]; then # do not do iterative dilation on subcort-only parcellation
+            
+            echo "${parc} parcellation intersection with GM; pcort is -- ${pcort}"
+
+            fileOut="${T1path}/T1_parc_${parc}_dil.nii.gz"
         
-        # Dilate the parcellation.
-        cmd="fslmaths ${fileIn} -dilD ${fileOut}"
-        log $cmd
-        eval $cmd 
-
-        # Iteratively mask the dilated parcellation with GM.
-        fileMul="${T1path}/T1_GM_mask.nii.gz"
-        checkisfile ${fileMul}
-
-        # # Apply subject GM mask
-        fileOut2="${T1path}/T1_GM_parc_${parc}.nii.gz"
-
-        cmd="fslmaths ${fileOut} -mul ${fileMul} ${fileOut2}"
-        log $cmd
-        eval $cmd 
-
-        # Dilate and remask to fill GM mask a set number of times
-        fileOut3="${T1path}/T1_GM_parc_${parc}_dil.nii.gz"
-
-        for ((j=1; j<=${configs_T1_numDilReMask}; j++)); do
-
-            cmd="fslmaths ${fileOut2} -dilD ${fileOut3}"
-            log $cmd
-            eval $cmd
-
-            cmd="fslmaths ${fileOut3} -mul ${fileMul} ${fileOut2}"
+            # Dilate the parcellation.
+            cmd="fslmaths ${fileIn} -dilD ${fileOut}"
             log $cmd
             eval $cmd 
 
-        done 
+            # Iteratively mask the dilated parcellation with GM.
+            fileMul="${T1path}/T1_GM_mask.nii.gz"
+            checkisfile ${fileMul}
 
-        # Remove the left over dil parcellation images.
-        cmd="rm ${fileOut} ${fileOut3}"
-        log $cmd
-        eval $cmd 
+            # # Apply subject GM mask
+            fileOut2="${T1path}/T1_GM_parc_${parc}.nii.gz"
+
+            cmd="fslmaths ${fileOut} -mul ${fileMul} ${fileOut2}"
+            log $cmd
+            eval $cmd 
+
+            # Dilate and remask to fill GM mask a set number of times
+            fileOut3="${T1path}/T1_GM_parc_${parc}_dil.nii.gz"
+
+            for ((j=1; j<=${configs_T1_numDilReMask}; j++)); do
+
+                cmd="fslmaths ${fileOut2} -dilD ${fileOut3}"
+                log $cmd
+                eval $cmd
+
+                cmd="fslmaths ${fileOut3} -mul ${fileMul} ${fileOut2}"
+                log $cmd
+                eval $cmd 
+
+            done 
+
+            # Remove the left over dil parcellation images.
+            cmd="rm ${fileOut} ${fileOut3}"
+            log $cmd
+            eval $cmd 
+        fi
+
+        # Intersect any subcortical parc with subcortical mask from FSL first.
+        if [[ "${psubcortonly}" -eq 0 ]]; then 
+
+            fileSubMask="${T1path}/T1_subcort_mask.nii.gz"
+            fileOut="${T1path}/T1_GM_parc_${parc}.nii.gz"
+
+            cmd="fslmaths ${fileIn} -dilD -mul ${fileSubMask} ${fileOut}"
+            log $cmd
+            eval $cmd 
+
+        fi
 
         if [ "${pcort}" -eq 1 ]; then
 
@@ -758,7 +773,7 @@ if ${flags_T1_parc}; then
 
                         if ! ${onesubcort} ; then
                             # check that parcellation is available in T1 space
-                            fileSubcortUser="T1_parc_${parcii}.nii.gz"
+                            fileSubcortUser="T1_GM_parc_${parcii}.nii.gz"
 
                             if [[ -f "${T1path}/${fileSubcortUser}" ]]; then
